+++ src/main/java/com/cg/flightmgmt/Controller/UserServiceController.java
@@ -18,6 +18,8 @@ import com.cg.flightmgmt.Service.UserService;
 @RestController
 public class UserServiceController {
 	
+	public static int logValidator = 0;
+	public static String UserType = "";
 	@Autowired
 	private UserService userService;
 	
@@ -33,31 +35,79 @@ public class UserServiceController {
 	@GetMapping("/viewbyId/{userId}")
 	public ResponseEntity<?> viewUser(@PathVariable("userId") BigInteger userId)
 	{
-		Users userdb=userService.viewUser(userId);
-		return ResponseEntity.ok(userdb);
+		if(logValidator == 1) 
+		{
+			if(UserType.equalsIgnoreCase("admin"))
+			{
+				Users userdb=userService.viewUser(userId);
+				return ResponseEntity.ok(userdb);
+			}
+			else
+				return ResponseEntity.ok("You don't have admin privileges");
+		}
+		else
+			return ResponseEntity.ok("You have not logged in yet");
 	}
 	
 	@GetMapping("/viewallUsers")
 	public ResponseEntity<?> viewUser()
 	{
-		List<Users> list= userService.viewUser();
-		return ResponseEntity.ok(list);
+		if(logValidator == 1) 
+		{
+			if(UserType.equalsIgnoreCase("admin"))
+			{
+				List<Users> list= userService.viewUser();
+				return ResponseEntity.ok(list);
+			}
+			else
+				return ResponseEntity.ok("You don't have admin privileges");
+		}
+		else
+			return ResponseEntity.ok("You have not logged in yet");
 	}
 	
 	@PutMapping("/updateUser")
 	public ResponseEntity<?> updateUser(@RequestBody Users user)
 	{
-		userService.validateUser(user);
-		Users userdb= userService.updateUser(user);
-		return ResponseEntity.ok(userdb);
-		
+		if(logValidator == 1) 
+		{
+			userService.validateUser(user);
+			Users userdb= userService.updateUser(user);
+			return ResponseEntity.ok(userdb);
+		}
+		else
+			return ResponseEntity.ok("You have not logged in yet");
 	}
-	
+		
 	@DeleteMapping("/deletebyId/{userId}")
 	public ResponseEntity<?> deleteUser(@PathVariable("userId") BigInteger userId)
 	{
-		userService.deleteUser(userId);
-		return ResponseEntity.ok("UserId"+userId+" has been deleted.");
+		if(logValidator == 1) 
+		{
+			userService.deleteUser(userId);
+			return ResponseEntity.ok("UserId"+userId+" has been deleted.");
+		}
+		else
+			return ResponseEntity.ok("You have not logged in yet");
 	}
-
+	@GetMapping("/Login/{userId}/{userPassword}")
+	public ResponseEntity<?> loggingUser(@PathVariable("userId") BigInteger userId,@PathVariable("userPassword") String userPassword)
+	{
+		boolean value = userService.loginDetails(userId, userPassword);
+		if(value == true)
+		{	
+			logValidator = 1;
+			UserType = userService.viewUser(userId).getUserType();
+			return ResponseEntity.ok("Logged In");
+		}
+		else
+			return ResponseEntity.ok("Invalid Credentials");
+	}
+	@GetMapping("/Logout")
+	public ResponseEntity<?> logOutUser()
+	{
+		logValidator = 0;
+		UserType = "";
+		return ResponseEntity.ok("Logged Out");
+	}
 }
